Auth has been half refactored, fixed provider fetching outside of widget trees!
It successfully compiles and runs, except when you try to fetch all of the chatrooms currently in use in `messaging.dart`.  Will probably have a document reference for the users internal list of chatrooms, it'll avoid duplicate setting and getting of documents!

Uploading it now so I have access to it on my laptop during the day!